used async & await to get game id
