+++ src/index.js
@@ -1,2 +1,3 @@
 import './style.css';
-import './modules/addScore.js';
+import './modules/LeaderboardAPI.js';
+
